Break up underlying functionality and Tango behaviors.
